refactor(chat_v2): extract pipeline hooks for approval and audit
WI-13: 定义 PipelineHook trait（before_turn / before_tool / after_tool /
before_compaction 四个真实调用点），把 execute_single_tool 内联的审批准入
（Kill Switch、allowlist、trusted automation、功能开关、命令守卫、审批作用域
绑定、AuthorityGate、ApprovalManager、重绑定复核、计划批准原子消费）迁为
内置 ApprovalGateHook::before_tool；external MCP 安全边界注记与 trusted
automation 预授权标记迁为 TaskAuditHook::after_tool。两个 hook 默认注册，
行为与迁移前等价。tool_loop.rs 5507 -> 4222 行（-23.3%）。
